refactor(model-core): split model requirement data
